🎨 Fix style of blog section
